% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_vesselhistory.R
\name{clean_vesselhistory}
\alias{clean_vesselhistory}
\title{Clean vesselhistory_raw data}
\usage{
clean_vesselhistory(vesselhistory_raw)
}
\arguments{
\item{vesselhistory_raw}{}
}
\value{
a tibble of cleaned vesselhistory data
}
\description{
This function cleans the vesselhistory_raw data.
}
\examples{
clean_vesselhistory(vesselhistory_raw)
}
