% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_vesselinfo.R
\name{clean_vesselinfo}
\alias{clean_vesselinfo}
\title{Clean vesselinfo_raw data}
\usage{
clean_vesselinfo(vesselinfo_raw)
}
\arguments{
\item{vesselinfo_raw}{}
}
\value{
a tibble of cleaned data
}
\description{
This function cleans the `vesselinfo_raw data`
}
\examples{
clean_vesselinfo(vesselinfo_raw)
}
