% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_weather_terminal.R
\name{clean_weather_terminal}
\alias{clean_weather_terminal}
\title{Clean weather_terminal_history_raw Data}
\usage{
clean_weather_terminal(weather_terminal_history_raw)
}
\arguments{
\item{weather_terminal_history_raw}{}
}
\value{
a tibble of cleaned data
}
\description{
This function cleans the `weather_terminal_history_raw data`
.
}
\examples{
clean_weather_terminal(weather_terminal_history_raw)
}
