% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_timestamp.R
\name{convert_timestamp}
\alias{convert_timestamp}
\title{Convert Timestamp}
\usage{
convert_timestamp(timestamp)
}
\arguments{
\item{timestamp}{The timestamp as a string.}
}
\value{
POSIXt human readable time stamp
}
\description{
The Seattle Ferry data contains many timestamps in the following
format: "/Date(1716553800000-0700)/". This function will convert
the timestamp into a readable date, time, timezone POSIXt object.
}
\examples{
convert_timestamp("/Date(1716553800000-0700)/")
}
