% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_terminal_weather_history.R
\name{get_terminal_weather_history}
\alias{get_terminal_weather_history}
\title{Get Terminal Weather History}
\usage{
get_terminal_weather_history(terminal, start_date, end_date, pause = 0)
}
\arguments{
\item{terminal}{Seattle ferry terminal name.}

\item{start_date}{Start date of history window.}

\item{end_date}{End date of history window.}

\item{pause}{pause is specified in number of seconds. Inject a delay if the API rejects requests for too many at once. One year of data for one location is equivalent to about 16 API calls and there is a limit of 600 per minute so for 20 terminals, add a pause.}
}
\value{
Tibble where each row is a terminal and weather information on a specific date.
}
\description{
Get Terminal Weather History
}
\examples{
get_terminal_weather_history("Anacortes", start_date = "2024-06-01", end_date = "2024-06-03", pause = 0)
}
