% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_terminalinfo.R
\name{get_terminalinfo}
\alias{get_terminalinfo}
\title{Get Terminal Information}
\usage{
get_terminalinfo()
}
\value{
A tibble.
}
\description{
A helper function to extract the name, ID, latitude, and longitude for
the various Seattle Ferry stations
}
\examples{
get_terminalinfo()
}
