% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vesselinfo.R
\name{get_vesselinfo}
\alias{get_vesselinfo}
\title{Get Vessel Info}
\usage{
get_vesselinfo(VesselID = NULL)
}
\arguments{
\item{VesselID}{Vessel ID number. Leave blank to extract all vessel info.}
}
\value{
Tibble of vessel information.
}
\description{
Get Vessel Info
}
\examples{
get_vesselinfo()
get_vesselinfo(1)
}
