% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vessel_accommodations.R
\name{vessel_accommodations}
\alias{vessel_accommodations}
\title{Vessel Accommodations}
\usage{
vessel_accommodations(
  VesselID = NULL,
  apiaccesscode = Sys.getenv("WSDOT_ACCESS_CODE")
)
}
\arguments{
\item{VesselID}{Numeric ID for a specific vessel. If blank, all vessel information will be returned.}

\item{apiaccesscode}{WSDOT Access Code.}
}
\value{
Vessel information organized as a list.
}
\description{
A wrapper to extract various information about a specific vessel.
}
\examples{
vessel_accommodations(38)
}
