% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vessel_basics.R
\name{vessel_basics}
\alias{vessel_basics}
\title{Vessel Basics}
\usage{
vessel_basics(VesselID = NULL, apiaccesscode = Sys.getenv("WSDOT_ACCESS_CODE"))
}
\arguments{
\item{VesselID}{Numeric ID for a specific vessel. If blank, all vessel information will be returned.}

\item{apiaccesscode}{WSDOT Access Code.}
}
\value{
Vessel basic information organized as a list.
}
\description{
A wrapper to extract various basic information about a specific vessel.
}
\examples{
vessel_basics(38)
}
