% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_df_to_db.R
\name{write_df_to_db}
\alias{write_df_to_db}
\title{A wrapper around DBI::dbWriteTable that Write a data frame to the specified database and schema}
\usage{
write_df_to_db(
  df,
  name,
  con,
  schema = Sys.getenv("DATABASE_SCHEMA"),
  append = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{df}{input databframe}

\item{name}{the name of the table to be written to the database}

\item{con}{the database connection}

\item{schema}{schema, defaults to the value of the environment variable DATABASE_SCHEMA}

\item{append}{argument to dbWriteTable}

\item{overwrite}{argument to dbWriteTable}
}
\description{
A wrapper around DBI::dbWriteTable that Write a data frame to the specified database and schema
}
\examples{
write_df_to_db(vesselhistory_raw, my_username, con, append = TRUE)
}
