% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_metadata.R
\name{write_metadata}
\alias{write_metadata}
\title{Write YAML metadata to facilitate conditional Quarto rendering}
\usage{
write_metadata(key, value)
}
\arguments{
\item{key}{A string representing the metadata key}

\item{value}{The metadata value. When used in conjunction with the Quarto chunk option `.content-visible when-meta` or `.content-hidden when-meta`, this should be a boolean TRUE or FALSE}
}
\value{
asis knitr output
}
\description{
This function writes dynamically-generated YAML in a Quarto document of the form:
---
key: value
---
It is used in conjunction with the Quarto chunk option `.content-visible when-meta` or `.content-hidden when-meta`
}
\examples{
write_meta("condition_1", TRUE)
}
